requirements.txt
annotated-types==0.5.0
black==23.9.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Flask==2.3.3
flask-swagger-ui==4.11.1
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==2.3.7
